move listenForJoins into room package
